refactor : utils 정리
